fix: rename file, extract helper
