+++ src/app/availability/[slug]/page.tsx
@@ -12,8 +12,11 @@ import {
     getTimeFromHourMinuteString,
     SAMPLE_MEMBERS,
 } from "@/lib/availability/utils";
-import { getExistingMeeting } from "@/lib/db/databaseUtils";
-import { getAvailability, getMeetingDates } from "@/lib/db/utils";
+import {
+    getAvailability,
+    getExistingMeeting,
+    getExistingMeetingDates,
+} from "@/lib/db/databaseUtils";
 import { HourMinuteString } from "@/lib/types/chrono";
 import { cn } from "@/lib/utils";
 
@@ -36,7 +39,7 @@ export default async function Page({ params }: PageProps) {
         redirect("/error");
     }
 
-    const meetingDates = await getMeetingDates(slug);
+    const meetingDates = await getExistingMeetingDates(slug);
     // const availability = user ? await getAvailability(user, slug) : [];
     const availability = await getAvailability({
         userId: "123", // TODO (#auth): replace with actual user from session
